feat: Phase 3 — Place Pool page UI rewrite + bug fixes
UI Rewrite:
- New CityPicker grouped dropdown (countries + cities + place counts + emoji flags)
- One scope state shared by all tabs (replaces breadcrumb bar + seeding dropdown)
- All RPCs rewired to FK-based params (city_id UUID + country_code)
- New RejectedTab with approve (category picker modal) + delete actions
- StaleTab removed, SeedingTabWrapper removed
- 6 tabs: Overview, Browse, Map, Seeding, Photos, Rejected

Bug Fixes:
- AI validation job now resolves UUID/code to text names for edge fn filters
- PhotoTab passes registeredCity.name/.country to backfill edge fn
- Category breakdown columns match new RPC return shape
- "Uncategorized" stat replaced with "Pending Review" (ai_pending_count)
- City picker z-index raised to 9999 (above Leaflet map)
- Rejected tab: names clickable → PlaceDetailModal with full data
- ROUND(double precision) → ROUND(::NUMERIC) fix in category + city RPCs

Migrations updated: 20260407300000_rpc_redesign_fk_based.sql (ROUND cast fix)
RPC fixes applied live. No mobile impact.

+++ supabase/migrations/20260407300000_rpc_redesign_fk_based.sql
@@ -155,7 +155,7 @@ BEGIN
     COUNT(DISTINCT pp.ai_categories[1]) FILTER (
       WHERE pp.is_active AND pp.ai_approved = true AND pp.ai_categories IS NOT NULL AND array_length(pp.ai_categories, 1) > 0
     )::INTEGER AS category_coverage,
-    ROUND(AVG(pp.rating) FILTER (WHERE pp.is_active AND pp.ai_approved = true AND pp.rating IS NOT NULL), 1) AS avg_rating
+    ROUND((AVG(pp.rating) FILTER (WHERE pp.is_active AND pp.ai_approved = true AND pp.rating IS NOT NULL))::NUMERIC, 1) AS avg_rating
   FROM seeding_cities sc
   LEFT JOIN place_pool pp ON pp.city_id = sc.id
   WHERE sc.country_code = p_country_code
@@ -196,7 +196,7 @@ BEGIN
         / COUNT(*)
       )::INTEGER ELSE 0
     END AS photo_pct,
-    ROUND(AVG(pp.rating) FILTER (WHERE pp.rating IS NOT NULL), 1) AS avg_rating
+    ROUND((AVG(pp.rating) FILTER (WHERE pp.rating IS NOT NULL))::NUMERIC, 1) AS avg_rating
   FROM place_pool pp
   LEFT JOIN seeding_cities sc ON pp.city_id = sc.id
   WHERE pp.is_active
